bugfix: do288 (OCP 4.2) lab ch10s03: todo-backend
The "todo-backend" application fails to start in the current
version of the do288 OCP 4.2 course, which prevents completion
of the lab "Containerizing and Deploying a Service" in chapter
10, section 3.

This "todo-backend" application's "package.json" file appears to
have been copied from a previous lab ("todo-single"), but the
version number of the "restify" depdency has been transposed from
"4.3.0" (working) to "4.0.3" (broken).

This commit enables this lab to be completed successfully.